aula e

+++ src/routes/memories.ts
@@ -3,8 +3,15 @@ import { prisma } from "../libs/prisma";
 import { z } from "zod";
 
 export async function memoriesRoutes(app: FastifyInstance) {
-    app.get("/memories", async () => {
+    app.addHook("preHandler", async (request) => {
+        await request.jwtVerify();
+    });
+
+    app.get("/memories", async (request) => {
         const memories = await prisma.memory.findMany({
+            where: {
+                userId: request.user.sub,
+            },
             orderBy: {
                 createdAt: "asc",
             },
@@ -33,14 +40,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
                 content,
                 coverUrl,
                 isPublic,
-                userId: "684255e3-7c15-4bfd-935c-1f2b160df97c",
+                userId: request.user.sub,
             },
         });
 
         return memory;
     });
 
-    app.get("/memories/:id", async (request) => {
+    app.get("/memories/:id", async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -51,10 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
             where: { id },
         });
 
+        if (!memory.isPublic && memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
+
         return memory;
     });
 
-    app.put("/memories/:id", async (request) => {
+    app.put("/memories/:id", async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
@@ -69,7 +80,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
         const { content, coverUrl, isPublic } = bodySchema.parse(request.body);
 
-        const memory = await prisma.memory.update({
+        let memory = await prisma.memory.findUniqueOrThrow({
+            where: { id },
+        });
+
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
+
+        memory = await prisma.memory.update({
             where: { id },
             data: {
                 content,
@@ -81,12 +100,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
         return memory;
     });
 
-    app.delete("/memories/:id", async (request) => {
+    app.delete("/memories/:id", async (request, reply) => {
         const paramsSchema = z.object({
             id: z.string().uuid(),
         });
 
         const { id } = paramsSchema.parse(request.params);
+
+        const memory = await prisma.memory.findUniqueOrThrow({
+            where: { id },
+        });
+
+        if (memory.userId !== request.user.sub) {
+            return reply.status(401).send();
+        }
 
         await prisma.memory.delete({
             where: { id },
